+++ apps/_root/ui/swap/widget/Widget.tsx
@@ -1,5 +1,3 @@
-'use client'
-
 import { Widget as UIWidget } from '@sushiswap/ui/future/components/widget'
 import React, { FC } from 'react'
 import { SwitchAppType } from './SwitchAppType'
@@ -13,10 +11,10 @@ import { useSwapState } from '../trade/TradeProvider'
 import { AppType } from '@sushiswap/ui'
 import { SwapButtonCrossChain } from './SwapButtonCrossChain'
 import { SettingsModule, SettingsOverlay } from '@sushiswap/ui/future/components/settings'
+import { ThunderCoreBanner } from './ThunderCoreBanner'
 
 export const Widget: FC = () => {
   const { appType } = useSwapState()
-
   return (
     <div className="flex flex-col gap-4">
       <WidgetTitleV2 />
@@ -35,7 +33,9 @@ export const Widget: FC = () => {
         <SwapCurrencyInput />
         <SwitchTokensButton />
         <SwapCurrencyOutput />
+        {/* <SwapButton /> */}
         {appType === AppType.Swap ? <SwapButton /> : <SwapButtonCrossChain />}
+        <ThunderCoreBanner />
       </UIWidget.Content>
     </div>
   )