+++ apps/analytics/components/TableSection/TableSection.tsx
@@ -1,61 +1,66 @@
 import { Tab } from '@headlessui/react'
-import React, { FC, Fragment, useCallback } from 'react'
+import React, { FC, Fragment } from 'react'
 
-import { PoolTable } from '../PoolTable'
-import { SelectedTable, usePoolFilters } from '../PoolsFiltersProvider'
-import { TableFilters } from '../Table/TableFilters'
-import { TokenTable } from '../TokenTable'
+import { Pools } from './Pools'
 import { Button } from '@sushiswap/ui/future/components/button'
+// import { TokenTable } from './Tokens'
+import Container from '@sushiswap/ui/future/components/Container'
+import { FuroTokenTable } from './FuroTokens'
 
 export const TableSection: FC = () => {
-  const { setFilters, selectedTable } = usePoolFilters()
-
-  const onChange = useCallback(
-    (val: number) => {
-      setFilters({
-        selectedTable: val === 0 ? SelectedTable.VerifiedPools : SelectedTable.UnverifiedPools,
-      })
-    },
-    [setFilters]
-  )
-
   return (
-    <section className="flex flex-col">
-      <Tab.Group onChange={onChange}>
-        <div className="flex items-center gap-2 mb-4">
-          <Tab as={Fragment}>
+    <section className='flex flex-col'>
+      <Tab.Group>
+        <Container maxWidth='7xl' className='px-4 mx-auto'>
+          <div className='flex items-center gap-2 mb-4'>
+            <Tab as={Fragment}>
+              {({ selected }) => (
+                <Button
+                  size='sm'
+                  variant={selected ? 'outlined' : 'empty'}
+                  color='default'
+                  className='!rounded-full !h-[36px]'
+                >
+                  SushiSwap 🍣
+                </Button>
+              )}
+            </Tab>
+            <Tab as={Fragment}>
+              {({ selected }) => (
+                <Button
+                  size='sm'
+                  variant={selected ? 'outlined' : 'empty'}
+                  color='default'
+                  className='!rounded-full !h-[36px]'
+                >
+                  SushiPay 💸
+                </Button>
+              )}
+            </Tab>
+            {/* <Tab as={Fragment}>
             {({ selected }) => (
-              <Button size="sm" variant={selected ? 'outlined' : 'empty'} color="default">
-                Verified Pools
+              <Button size="sm" variant={selected ? 'outlined' : 'empty'} color="default"    className="!rounded-full !h-[36px]">
+                Sushi Vault
               </Button>
             )}
-          </Tab>
-          <Tab as={Fragment}>
+          </Tab> */}
+
+            {/* <Tab as={Fragment}>
             {({ selected }) => (
-              <Button size="sm" variant={selected ? 'outlined' : 'empty'} color="default">
-                Unverified Pools
-              </Button>
-            )}
-          </Tab>
-          {/* <Tab as={Fragment}>
-            {({ selected }) => (
-              <Button size="sm" variant={selected ? 'outlined' : 'empty'} color="default">
+              <Button size="sm" variant={selected ? 'outlined' : 'empty'} color="default" >
                 Tokens
               </Button>
             )}
           </Tab> */}
-        </div>
-        <TableFilters />
-        <Tab.Panels>
+          </div>
+        </Container>
+        <Tab.Panels className='bg-gray-50 dark:bg-white/[0.02] pt-4'>
           <Tab.Panel unmount={false}>
-            <PoolTable isWhitelisted={true} />
+            <Pools />
           </Tab.Panel>
           <Tab.Panel unmount={false}>
-            <PoolTable isWhitelisted={false} />
+            <FuroTokenTable />
           </Tab.Panel>
-          {/* <Tab.Panel unmount={false}>
-            <TokenTable />
-          </Tab.Panel> */}
         </Tab.Panels>
       </Tab.Group>
     </section>