+++ apps/earn/components/AddPage/AddSectionReviewModalTrident.tsx
@@ -73,7 +73,7 @@ export const AddSectionReviewModalTrident: FC<AddSectionReviewModalTridentProps>
   const contract = useTridentRouterContract(chainId)
   const [slippageTolerance] = useSlippageTolerance('addLiquidity')
   const slippagePercent = useMemo(() => {
-    return new Percent(Math.floor(+slippageTolerance * 100), 10_000)
+    return new Percent(Math.floor(+(slippageTolerance === 'AUTO' ? '0.5' : slippageTolerance) * 100), 10_000)
   }, [slippageTolerance])
 
   const [minAmount0, minAmount1] = useMemo(() => {