+++ apps/earn/components/RewardsSection/Tables/RewardsTableV3/Cells/index.ts
@@ -1,0 +1,4 @@
+export * from './columns'
+export * from './RewardsV3NameCell'
+export * from './types'
+export * from './RewardsV3ClaimableCell'