+++ apps/earn/components/RewardsSection/Tables/RewardsTableV3/Cells/types.ts
@@ -1,0 +1,5 @@
+import { AngleRewardsPool } from '@sushiswap/react-query'
+
+export interface RewardTableV3CellProps {
+  row: AngleRewardsPool
+}