+++ apps/earn/lib/hooks/api/useGraphPools.ts
@@ -3,7 +3,6 @@ import { useMemo } from 'react'
 import useSWR from 'swr'
 import { getGraphPools } from '../../api'
 import { Protocol } from '@sushiswap/client'
-import { useEffectDebugger } from '@sushiswap/hooks'
 
 function transformGraphPool(graphPool: Awaited<ReturnType<typeof getGraphPools>>[0]): Pool {
   let protocol: Protocol = Protocol.SUSHISWAP_V2
@@ -81,17 +80,19 @@ export function getGraphPoolsUrl(poolIds: string[]) {
 // Returns the Pools type, to make everyone's life easier
 // Updates possibly stale values with subgraph values
 export const useGraphPools = (poolIds: string[]): Pools => {
-  const { data: graphPools, isLoading: isGraphPoolsLoading } = useSWR<Awaited<ReturnType<typeof getGraphPools>>>(
+  const {
+    data: graphPools,
+    isLoading: isGraphPoolsLoading,
+    error: graphPoolsError,
+  } = useSWR<Awaited<ReturnType<typeof getGraphPools>>>(
     poolIds.length > 0 ? getGraphPoolsUrl(poolIds) : null,
-    async (url) => fetch(url).then((data) => data.json())
+    async () => getGraphPools(poolIds)
   )
 
-  const { data: pools, isLoading: isPoolsLoading } = usePools({ args: { ids: poolIds } })
-
-  useEffectDebugger(() => {}, [graphPools, isGraphPoolsLoading, isPoolsLoading, poolIds, pools])
+  const { data: pools, isLoading: isPoolsLoading, error: poolsError } = usePools({ args: { ids: poolIds } })
 
   return useMemo(() => {
-    if (isGraphPoolsLoading || isPoolsLoading) {
+    if ((isGraphPoolsLoading && !graphPoolsError) || (isPoolsLoading && !poolsError)) {
       return []
     }