+++ apps/furo/components/vesting/CreateMultipleForm/CreateMultipleForm.tsx
@@ -44,14 +44,14 @@ export const CreateMultipleForm: FC<{ chainId: FuroVestingRouterChainId }> = ({ 
     <div className={classNames('flex flex-col gap-10')}>
       <Link href="/vesting/create" passHref={true} legacyBehavior>
         <a>
-          <button className="flex gap-3 font-medium group hover:text-white text-slate-200">
+          <button type="button" className="flex gap-3 font-medium group hover:text-white text-slate-200">
             <ArrowCircleLeftIcon width={24} height={24} /> <span>Create Vesting</span>
           </button>
         </a>
       </Link>
       <FormProvider {...methods}>
         <div className="flex flex-col gap-14">
-          <ImportErrorProvider<CreateMultipleVestingFormSchemaType>>
+          <ImportErrorProvider>
             <ImportZoneSection chainId={chainId} />
             <div className="w-full border-b border-slate-200/5" />
             <div className={review ? 'hidden' : ''}>