+++ packages/wagmi/src/future/contracts/actions/getStablePoolFactoryContract.ts
@@ -1,0 +1,7 @@
+import { stablePoolFactoryAbi, stablePoolFactoryAddress, StablePoolFactoryChainId } from '@sushiswap/trident-core'
+import { Address } from 'wagmi'
+
+export const getStablePoolFactoryContract = (chainId: number | undefined) => ({
+  address: (stablePoolFactoryAddress?.[chainId as StablePoolFactoryChainId] ?? '') as Address,
+  abi: stablePoolFactoryAbi?.[chainId as StablePoolFactoryChainId] ?? [],
+})